- Added rainbow sonar (left and right click to use)

+++ Source/Life/Source/Player.h
@@ -33,7 +33,7 @@ public:
 
 	vector<Sonar*> sonarList;
 
-	int sonarCooldown, specialCooldown;
+	float sonarCooldown, specialCooldown;
 	float sonarTimer, specialTimer;
 
 	float specialDuration, specialROF, specialTimer2, specialCounter;
